pyproject.toml
[project]
name = "simple-fullstack-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.3.0",
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "python-dotenv>=1.0.1",
]
